python-sdks/respan-exporter-haystack/pyproject.toml
[tool.poetry]
name = "respan-exporter-haystack"
version = "1.0.0"
description = "Respan integration for Haystack pipelines with tracing and logging support"
authors = ["Respan <team@respan.ai>"]
readme = "README.md"
license = "MIT"
packages = [{include = "respan_exporter_haystack", from = "src"}]
homepage = "https://respan.ai"
repository = "https://github.com/respan-ai/respan-sdks"
documentation = "https://docs.respan.ai"
keywords = ["haystack", "respan", "llm", "observability", "tracing", "monitoring"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
requests = "^2.32.5"
haystack-ai = "^2.24.1"
respan-sdk = "^2.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
python-dotenv = "^1.2.1"
sentence-transformers = "^5.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
